requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-wtf==1.2.1
supabase==2.5.0
python-dotenv==1.0.1
email-validator==2.1.1
